Extract header
