+++ src/app/inventory/PhoneStores.tsx
@@ -4,12 +4,10 @@ import StoreStats from 'app/store-stats/StoreStats';
 import clsx from 'clsx';
 import React, { useRef, useState } from 'react';
 import Hammer from 'react-hammerjs';
-import { Frame, Track, View, ViewPager } from 'react-view-pager';
-import StoreHeading from '../character-tile/StoreHeading';
 import ScrollClassDiv from '../dim-ui/ScrollClassDiv';
-import { hideItemPopup } from '../item-popup/item-popup';
 import { storeBackgroundColor } from '../shell/filters';
 import { InventoryBuckets } from './inventory-buckets';
+import PhoneStoresHeader from './PhoneStoresHeader';
 import { DimStore } from './store-types';
 import { StoreBuckets } from './StoreBuckets';
 import { getCurrentStore, getStore, getVault } from './stores-helpers';
@@ -38,11 +36,6 @@ export default function PhoneStores({ stores, buckets }: Props) {
   }
 
   const selectedStore = selectedStoreId ? getStore(stores, selectedStoreId)! : currentStore;
-
-  const onViewChange = (indices: number[]) => {
-    setSelectedStoreId(stores[indices[0]].id);
-    hideItemPopup();
-  };
 
   const handleSwipe: HammerListener = (e) => {
     const selectedStoreIndex = selectedStoreId
@@ -77,27 +70,12 @@ export default function PhoneStores({ stores, buckets }: Props) {
         style={storeBackgroundColor(selectedStore, 0, true, true)}
         onTouchStart={(e) => e.stopPropagation()}
       >
-        <ViewPager>
-          <Frame className="frame" autoSize={false}>
-            <Track
-              currentView={selectedStore.id}
-              contain={false}
-              onViewChange={onViewChange}
-              className="track"
-            >
-              {stores.map((store) => (
-                <View className="store-cell" key={store.id}>
-                  <StoreHeading
-                    store={store}
-                    selectedStore={selectedStore}
-                    onTapped={setSelectedStoreId}
-                    loadoutMenuRef={detachedLoadoutMenu}
-                  />
-                </View>
-              ))}
-            </Track>
-          </Frame>
-        </ViewPager>
+        <PhoneStoresHeader
+          selectedStore={selectedStore}
+          stores={stores}
+          loadoutMenuRef={detachedLoadoutMenu}
+          setSelectedStoreId={setSelectedStoreId}
+        />
       </ScrollClassDiv>
 
       <div className="detached" ref={detachedLoadoutMenu} />
